+++ src/prefect/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to
 - Add usage indicator u14
 - Add extract indicator e5
 - Add extract indicator e1
+- Add extract indicators e2 and e3
 
 #### Quality